requirements.txt
Django==4.2.7
django-filter==23.2
python-dotenv==1.0.0
